+++ src/main/java/asdf/lol/rofl/EFrameWorkUserIsIdiotException.java
@@ -1,3 +1,4 @@
+package asdf.lol.rofl;
 public class EFrameWorkUserIsIdiotException extends Exception {
 	public EFrameWorkUserIsIdiotException(String message) {
 		super(message);
