--- src/main/java/at/grueneis/spengergasse/lesson_plan/domain/BasePersistable.java
@@ -1,52 +1,0 @@
-/*
- * Joachim Grüneis
- * Copyright (C) 2013
- * All rights reserved.
- */
-package at.grueneis.spengergasse.lesson_plan.domain;
-
-/**
- * A common base class for persistable domain classes that holds the id
- * attribute and implements equals, hashCode and toString using the id attribute
- * only.
- */
-public class BasePersistable {
-
-    private Long id;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        BasePersistable that = (BasePersistable) o;
-
-        if (id != null ? !id.equals(that.id) : that.id != null) {
-            return false;
-        }
-
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        return id != null ? id.hashCode() : 0;
-    }
-
-    @Override
-    public String toString() {
-        return getClass().getSimpleName() + "[id=" + id + "]";
-    }
-}
