--- src/main/java/at/grueneis/spengergasse/lesson_plan/domain/LessonPlan.java
@@ -1,34 +1,0 @@
-/*
- * Joachim Grüneis
- * Copyright (C) 2013
- * All rights reserved.
- */
-package at.grueneis.spengergasse.lesson_plan.domain;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-
-public class LessonPlan {
-    private final String name;
-    private final SchoolYear schoolYear;
-    private final List<Lesson> lessons;
-
-    public LessonPlan(String name, SchoolYear schoolYear,
-            Collection<Lesson> lessons) {
-        this.name = name;
-        this.schoolYear = schoolYear;
-        this.lessons = new ArrayList<>();
-        for (Lesson lesson : lessons)
-            this.lessons.add(lesson);
-    }
-
-    public Collection<Lesson> getLessons() {
-        // Collection<Lesson> returnList = new ArrayList<>();
-        // for (Lesson lesson : this.lessons)
-        // returnList.add(lesson);
-        // return returnList;
-        return Collections.unmodifiableCollection(lessons);
-    }
-}
