--- src/main/java/at/grueneis/spengergasse/lesson_plan/domain/SchoolYear.java
@@ -1,75 +1,0 @@
-/*
- * Joachim Grüneis
- * Copyright (C) 2013
- * All rights reserved.
- */
-package at.grueneis.spengergasse.lesson_plan.domain;
-
-import java.util.Date;
-
-public class SchoolYear {
-    private final String name;
-    private final Date start;
-    private final Date end;
-
-    public SchoolYear(String name, Date start, Date end) {
-        this.name = name;
-        this.start = start;
-        this.end = end;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Date getStart() {
-        return (Date) start.clone();
-    }
-
-    public Date getEnd() {
-        return (Date) end.clone();
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((end == null) ? 0 : end.hashCode());
-        result = prime * result + ((name == null) ? 0 : name.hashCode());
-        result = prime * result + ((start == null) ? 0 : start.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        SchoolYear other = (SchoolYear) obj;
-        if (end == null) {
-            if (other.end != null)
-                return false;
-        } else if (!end.equals(other.end))
-            return false;
-        if (name == null) {
-            if (other.name != null)
-                return false;
-        } else if (!name.equals(other.name))
-            return false;
-        if (start == null) {
-            if (other.start != null)
-                return false;
-        } else if (!start.equals(other.start))
-            return false;
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "SchoolYear [name=" + name + ", start=" + start + ", end=" + end
-                + "]";
-    }
-}
